pyproject.toml
[project]
name = "test-case-task-manager"
version = "0.1.0"
description = ""
authors = [
    {name = "shelpyakovstepan",email = "step3210shelpyakov@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[all] (>=0.116.1,<0.117.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic-postgresql-enum (>=1.8.0,<2.0.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
    "pyright (>=1.1.403,<2.0.0)",
    "ruff (>=0.12.9,<0.13.0)",
    "isort (>=6.0.1,<7.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "uuid (>=1.30,<2.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pytest-cov (>=6.2.1,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
